Pull out GithubApiError and GithubRateLimitedError to a separate file
