+++ extensions/ql-vscode/src/extension.ts
@@ -53,9 +53,11 @@ import {
   DistributionUpdateCheckResultKind,
   FindDistributionResult,
   FindDistributionResultKind,
+} from "./codeql-cli/distribution";
+import {
   GithubApiError,
   GithubRateLimitedError,
-} from "./codeql-cli/distribution";
+} from "./codeql-cli/github-api-error";
 import { tmpDir, tmpDirDisposal } from "./tmp-dir";
 import { prepareCodeTour } from "./code-tour/code-tour";
 import {
